Restructure project into gw-siren-pipeline package
