+++ gw-siren-pipeline/README.md
@@ -1,0 +1,14 @@
+# GW Siren Pipeline
+
+A Python project for gravitational wave data analysis and H0 estimation.
+
+## Project Structure (Initial)
+
+- `gwsiren/`: Core library package.
+- `scripts/`: Executable analysis scripts (to be populated).
+- `tests/`: Unit, integration, and end-to-end tests.
+- Other `.py` files: Modules and scripts to be refactored.
+
+## Setup (Placeholder)
+
+To be updated as the project is refactored.
