+++ gw-siren-pipeline/config.yaml
