+++ gw-siren-pipeline/gwsiren/config.py
