+++ gw-siren-pipeline/gwsiren/data/catalogs.py
