+++ gw-siren-pipeline/sample_path.py
