+++ gw-siren-pipeline/tests/test_config.py
