+++ gw-siren-pipeline/tests/unit/test_event_data_extractor.py
